+++ src/main/java/edu/miu/annapurnabe/api/v1/TestResource.java
@@ -1,4 +1,4 @@
-package edu.miu.annapurnabe.api;
+package edu.miu.annapurnabe.api.v1;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
